pyproject.toml
[project]
name = "tl10k"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "edgartools>=3.11.1",
    "openai>=1.61.1",
    "playwright>=1.46.0",
    "requests>=2.32.3",
    "rich>=13.9.4",
]
